datatypes checked

+++ src/models/users.rs
@@ -28,7 +28,7 @@ pub struct UserSkill {
 	pub user_id: uuid::Uuid,
 	pub skill_id: uuid::Uuid,
 	pub skillscopelevel_id: uuid::Uuid,
-	pub years: Option<f32>,
+	pub years: Option<f64>,
 	pub updated_by: String,
 }
